+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## master (unreleased)
+
+- Fix issue with `auto_preload` tries to preload association by attribute name
+  instead of by `method` when `method` provided
+
 ## [0.3.1] - 2025-08-19
 
 - Fix issue with `auto_preload` tries to preload associations for attributes